'Create directory for shift_net models. It started to be unclean'
