+++ models/shift_net/base_model.py
