+++ models/shift_net/shiftnet_model.py
@@ -1,35 +1,21 @@
-import numpy as np
 import torch
-import os
-import util.util as util
-from PIL import Image
 from torch.nn import functional as F
-from .base_model import BaseModel
-from . import networks
+
+import util.util as util
+from models import networks
+from models.shift_net.base_model import BaseModel
+
 
 class ShiftNetModel(BaseModel):
     def name(self):
         return 'ShiftNetModel'
-
-
-    def create_random_mask(self):
-        if self.mask_type == 'random':
-            if self.opt.mask_sub_type == 'fractal':
-                mask = util.create_walking_mask ()  # create an initial random mask.
-
-            elif self.opt.mask_sub_type == 'rect':
-                mask = util.create_rand_mask ()
-
-            elif self.opt.mask_sub_type == 'island':
-                mask = util.wrapper_gmask (self.opt)
-        return mask
 
     def initialize(self, opt):
         BaseModel.initialize(self, opt)
         self.opt = opt
         self.isTrain = opt.isTrain
         # specify the training losses you want to print out. The program will call base_model.get_current_losses
-        self.loss_names = ['G_GAN', 'G_L1', 'D']
+        self.loss_names = ['G_GAN', 'G_L1', 'D_real', 'D_fake']
         # specify the images you want to save/display. The program will call base_model.get_current_visuals
         self.visual_names = ['real_A', 'fake_B', 'real_B']
         # specify the models you want to save to the disk. The program will call base_model.save_networks and base_model.load_networks
@@ -55,7 +41,24 @@ class ShiftNetModel(BaseModel):
             assert opt.fixed_mask == 1, "Center mask must be fixed mask!"
 
         if self.mask_type == 'random':
-            self.create_random_mask()
+            res = 0.06 # the lower it is, the more continuous the output will be. 0.01 is too small and 0.1 is too large
+            density = 0.25
+            MAX_SIZE = 10000
+            maxPartition = 30
+            low_pattern = torch.rand(1, 1, int(res*MAX_SIZE), int(res*MAX_SIZE)).mul(255)
+            pattern = F.interpolate(low_pattern, (MAX_SIZE, MAX_SIZE), mode='bilinear').detach()
+            low_pattern = None
+            pattern.div_(255)
+            pattern = torch.lt(pattern,density).byte()  # 25% 1s and 75% 0s
+            pattern = torch.squeeze(pattern).byte()
+            print('...Random pattern generated')
+            self.gMask_opts['pattern'] = pattern
+            self.gMask_opts['MAX_SIZE'] = MAX_SIZE
+            self.gMask_opts['fineSize'] = opt.fineSize
+            self.gMask_opts['maxPartition'] = maxPartition
+            self.gMask_opts['mask_global'] = self.mask_global
+            self.mask_global = util.create_gMask(self.gMask_opts) # create an initial random mask.
+
 
         self.wgan_gp = False
         # added for wgan-gp
@@ -125,11 +128,11 @@ class ShiftNetModel(BaseModel):
                 self.mask_global[:, :, int(self.opt.fineSize/4) + self.opt.overlap : int(self.opt.fineSize/2) + int(self.opt.fineSize/4) - self.opt.overlap,\
                                     int(self.opt.fineSize/4) + self.opt.overlap: int(self.opt.fineSize/2) + int(self.opt.fineSize/4) - self.opt.overlap] = 1
             elif self.opt.mask_type == 'random':
-                self.mask_global = self.create_random_mask().type_as(self.mask_global)
+                self.mask_global = util.create_gMask(self.gMask_opts).type_as(self.mask_global)
             else:
                 raise ValueError("Mask_type [%s] not recognized." % self.opt.mask_type)
         else:
-            self.mask_global = self.create_random_mask().type_as(self.mask_global)
+            self.mask_global = util.create_gMask(self.gMask_opts).type_as(self.mask_global)
 
         self.set_latent_mask(self.mask_global, 3, self.opt.threshold)
 
@@ -140,22 +143,6 @@ class ShiftNetModel(BaseModel):
         self.real_A = real_A
         self.real_B = real_B
         self.image_paths = input['A_paths']
-        
-    def set_input_with_mask(self, input, mask):
-        real_A = input['A'].to(self.device)
-        real_B = input['B'].to(self.device)
-        
-        self.mask_global = mask
-
-        self.set_latent_mask(mask, 3, self.opt.threshold)
-
-        real_A.narrow(1,0,1).masked_fill_(mask, 2*123.0/255.0 - 1.0)
-        real_A.narrow(1,1,1).masked_fill_(mask, 2*104.0/255.0 - 1.0)
-        real_A.narrow(1,2,1).masked_fill_(self.mask_global, 2*117.0/255.0 - 1.0)
-
-        self.real_A = real_A
-        self.real_B = real_B
-        self.image_paths = input['A_paths']       
 
     def set_latent_mask(self, mask_global, layer_to_last, threshold):
         self.ng_shift_list[0].set_mask(mask_global, layer_to_last, threshold)
@@ -195,20 +182,9 @@ class ShiftNetModel(BaseModel):
 
             self.loss_D = self.loss_D_fake - self.loss_D_real + gradient_penalty
         else:
-            self.pred_fake = self.netD(fake_AB.detach())
-
-            if self.opt.gan_type in ['vanilla', 'lsgan']:
-                self.loss_D_fake = self.criterionGAN(self.pred_fake, False)
-                self.loss_D_real = self.criterionGAN (self.pred_real, True)
-
-                self.loss_D = (self.loss_D_fake + self.loss_D_real) * 0.5
-
-            elif self.opt.gan_type == 're_s_gan':
-                self.loss_D = self.criterionGAN (self.pred_real - self.pred_fake, True)
-
-            elif self.opt.gan_type == 're_avg_gan':
-                self.loss_D =  (self.criterionGAN (self.pred_real - torch.mean(self.pred_fake), True) \
-                               + self.criterionGAN (self.pred_fake - torch.mean(self.pred_real), False)) / 2.
+            self.loss_D_fake = self.criterionGAN(self.pred_fake, False)
+            self.loss_D_real = self.criterionGAN(self.pred_real, True)
+            self.loss_D = (self.loss_D_fake + self.loss_D_real) * 0.5
 
         self.loss_D.backward()
 
@@ -216,27 +192,18 @@ class ShiftNetModel(BaseModel):
         # First, G(A) should fake the discriminator
         fake_AB = self.fake_B
         pred_fake = self.netD(fake_AB)
-
         if self.wgan_gp:
             self.loss_G_GAN = torch.mean(pred_fake)
         else:
-            if self.opt.gan_type in ['vanilla', 'lsgan']:
-                self.loss_G_GAN = self.criterionGAN(pred_fake, True)
+            self.loss_G_GAN = self.criterionGAN(pred_fake, True)
 
-            elif self.opt.gan_type == 're_s_gan':
-                pred_real = self.netD (self.real_B)
-                self.loss_G_GAN = self.criterionGAN (pred_fake - pred_real, True)
-
-            elif self.opt.gan_type == 're_avg_gan':
-                self.pred_real = self.netD(self.real_B)
-                self.loss_G_GAN =  (self.criterionGAN (self.pred_real - torch.mean(self.pred_fake), False) \
-                               + self.criterionGAN (self.pred_fake - torch.mean(self.pred_real), True)) / 2.
-
-            # Second, G(A) = B
+        # Second, G(A) = B
         self.loss_G_L1 = self.criterionL1(self.fake_B, self.real_B) * self.opt.lambda_A
 
-        self.loss_G = self.loss_G_L1 - self.loss_G_GAN * self.opt.gan_weight
-
+        if self.wgan_gp:
+            self.loss_G = self.loss_G_L1 - self.loss_G_GAN * self.opt.gan_weight
+        else:
+            self.loss_G = self.loss_G_L1 + self.loss_G_GAN * self.opt.gan_weight
 
         # Third add additional netG contraint loss!
         self.ng_loss_value = 0
