+++ models/soft_shift_net/innerSoftShiftTriple.py
