+++ models/soft_shift_net/innerSoftShiftTripleModule.py
